Refactored Code to Use Modern React Components
