+++ src/component/tile.js
@@ -1,10 +1,10 @@
-import {Component} from 'react'
-
-
-export class Tile extends Component {
-    render() {
-        return (
-            <div className={`bg-cover bg-center h-auto ${this.props.span == 2? 'col-span-2': 'col-span-3'}`} style={{backgroundImage: `url(https://picsum.photos/seed/${encodeURI(this.props.seed + '-' + this.props.id)}/1000?1)`}}></div>
-        )
-    }
+const Tile = ({ id, seed, span }) => {
+    return (
+        <div
+            className={`bg-cover bg-center h-auto ${span == 2? 'col-span-2': 'col-span-3'}`}
+            style={{backgroundImage: `url(https://picsum.photos/seed/${encodeURI(seed + '-' + id)}/1000?1)`}}>
+        </div>
+    );
 }
+
+export default Tile;
