+++ src/pages/_app.js
@@ -1,5 +1,7 @@
 import '../styles/index.css'
 
-export default function MyApp({ Component, pageProps }) {
+const MyApp = ({ Component, pageProps }) => {
   return <Component {...pageProps} />
 }
+
+export default MyApp;
